Refactor authentication and email verification routes; update Docker container name and PHP version
