+++ src/api/composer.json
@@ -20,7 +20,8 @@
   },
   "config": {
     "platform": {
-      "php": "8.4.0"
+      "php": "8.4.16",
+      "ext-redis": "6.1.0"
     },
     "sort-packages": true,
     "optimize-autoloader": true
