+++ src/api/configs/routes/parts/parts.php
@@ -2,6 +2,7 @@
 
 declare(strict_types=1);
 
-require_once __DIR__ . '/./auth.php';
-require_once __DIR__ . '/./user.php';
-require_once __DIR__ . '/./verifyEmail.php';
+require_once __DIR__ . '/./web/parts.php';
+require_once __DIR__ . '/./web/parts/auth.php';
+require_once __DIR__ . '/./web/parts/user.php';
+require_once __DIR__ . '/./web/parts/verifyEmail.php';
