+++ src/api/configs/routes/parts/web/parts/verifyEmail.php
@@ -3,13 +3,13 @@
 declare(strict_types=1);
 
 use Slim\Routing\RouteCollectorProxy;
-use JR\Tracker\Controller\VerifyEmailController;
+use JR\Tracker\Controller\Web\VerifyEmailController;
 use JR\Tracker\Middleware\ValidateSignatureMiddleware;
 
 function getVerifyEmailRoutes(RouteCollectorProxy $api)
 {
-    $api->group('/verifyEmail', function (RouteCollectorProxy $auth) {
-        $auth->post('/verify/{uuid}/{hash}', [VerifyEmailController::class, "verify"])
+    $api->group('/verifyEmail', function (RouteCollectorProxy $verifyEmail) {
+        $verifyEmail->post('/verify/{uuid}/{hash}', [VerifyEmailController::class, "verify"])
             ->setName('verify')
             ->add(ValidateSignatureMiddleware::class);
     });
