+++ src/api/configs/routes/web.php
@@ -7,9 +7,7 @@ use Slim\Routing\RouteCollectorProxy;
 
 return function (App $app) {
     $app->group('/api', function (RouteCollectorProxy $api) {
-        getAuthRoutes($api);
-        getUserRoutes($api);
-        getVerifyEmailRoutes($api);
+        getWebRoutes($api);
     });
 
     // TODO: Na endpointy, ktere se volaji jenom kdy6 jsem prihlaseny pridat VerifyEmailMidlleware a upravit authMiddleware a napojit asi na stejne endpointy
